% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_function.R
\name{coordinate_function}
\alias{coordinate_function}
\title{Longitude, Latitude, Date Conversion
'coordinate_function' returns latitude and longitude inputs into decimal degrees.
This function converts decimal minutes into decimal degrees for latitude and longitude data. The function also converts dates into R format.}
\usage{
coordinate_function(long, lat, date)
}
\arguments{
\item{long}{longitude coordinate}

\item{lat}{latitude coordinate}

\item{date}{(optional) date in DD-MMM-YY format}
}
\description{
Longitude, Latitude, Date Conversion
'coordinate_function' returns latitude and longitude inputs into decimal degrees.
This function converts decimal minutes into decimal degrees for latitude and longitude data. The function also converts dates into R format.
}
